+++ spec/runtime/globals_tracker_spec.rb
@@ -4,7 +4,7 @@ require_relative '../helper'
 
 # just to stop RuboCop from complaining about missing parameters
 def eval_function(text)
-  eval(text, binding, __FILE__, __LINE__ + 1)
+  eval(text, binding, File.join(__dir__, 'template'), __LINE__ + 1)
 end
 
 describe Bade::Runtime::GlobalsTracker do
@@ -45,10 +45,15 @@ describe Bade::Runtime::GlobalsTracker do
           NEW_GLOBAL_CONSTANT1 = 'abc'
 
           class Test1; end
+
+          module Abc
+            class AbcClass
+            end
+          end
         RB
       end
 
-      expect(@sut.caught_constants.map { |(_, name)| name }).to contain_exactly :NEW_GLOBAL_CONSTANT1, :Test1
+      expect(@sut.caught_constants.map { |(_, name)| name }).to contain_exactly :NEW_GLOBAL_CONSTANT1, :Test1, :Abc
     end
 
     it 'can remove caught constants' do
@@ -56,7 +61,9 @@ describe Bade::Runtime::GlobalsTracker do
         eval_function(<<-RB)
           NEW_GLOBAL_CONSTANT2 = 'abc'
 
-          class Test2; end
+          class Test2
+            def abc; end
+          end
         RB
       end
 
@@ -70,6 +77,20 @@ describe Bade::Runtime::GlobalsTracker do
       expect do
         eval_function('Test2')
       end.to raise_error(NameError)
+    end
+
+    it 'will not remove constants from required files' do
+      @sut.catch do
+        eval_function('require_relative "imported"')
+      end
+
+      expect($LOADED_FEATURES).to include(File.join(__dir__, 'imported.rb'))
+      expect(eval_function('ImportedModuleAbc.upcase("abc")')).to eq 'ABC'
+
+      @sut.clear_all
+
+      expect($LOADED_FEATURES).to include(File.join(__dir__, 'imported.rb'))
+      expect(eval_function('ImportedModuleAbc.upcase("abc")')).to eq 'ABC'
     end
   end
 end